Remove acl constants references
