+++ registry/src/registry/services/access_control_service.py
@@ -6,7 +6,8 @@ from beanie import PydanticObjectId
 from fastapi import HTTPException
 from fastapi import status as http_status
 
-from registry.core.acl_constants import PermissionBits, PrincipalType
+from registry_pkgs.models._generated import PrincipalType
+from registry_pkgs.models.enums import PermissionBits
 from registry.schemas.acl_schema import PermissionPrincipalOut, ResourcePermissions
 from registry.services.group_service import group_service
 from registry.services.user_service import user_service
