+++ registry/tests/unit/services/test_access_control_service.py
@@ -4,7 +4,8 @@ import pytest
 from beanie import PydanticObjectId
 from fastapi import HTTPException
 
-from registry.core.acl_constants import PermissionBits, ResourceType, RoleBits
+from registry_pkgs.models._generated import ResourceType
+from registry_pkgs.models.enums import PermissionBits, RoleBits
 from registry.schemas.acl_schema import ResourcePermissions
 from registry.services.access_control_service import ACLService
